dung-creating basic pdf-generator for [Inventory]

+++ back-end/src/main/java/com/dnth_underdog_241/online_fashion_shopping/util/ImportCostCalculatorUtil.java
@@ -44,7 +44,7 @@ public class ImportCostCalculatorUtil
      * @param quantity     the number of units being imported.
      * @return the total cost including shipping and taxes.
      */
-    public static double calculateTotalCost(double unitCost, int quantity)
+    public static double calculateTotalCost(Double unitCost, Long quantity)
     {
         double productCost = unitCost * quantity;
         double taxRate = generateRandomTaxRate();
